Updated the look
Updated the HTML by getting rid of unused tags and continued cleaning up. Updated the CSS margins and making the side nav fixed. Also started working on the Light/Dark mode. Updated the JS so that the date doesn't show the current day on all the notes including the past notes.